+++ src/setupTests.js
@@ -4,28 +4,3 @@
 // expect(element).toHaveTextContent(/react/i)
 // learn more: https://github.com/testing-library/jest-dom
 import '@testing-library/jest-dom/extend-expect';
-
-import React from 'react';
-import { render as rtlRender } from '@testing-library/react';
-import { createStore } from 'redux';
-import { Provider } from 'react-redux';
-import reducer from './reducers';
-
-function render(
-  ui,
-  {
-    initialState,
-    store = createStore(reducer, initialState),
-    ...renderOptions
-  } = {},
-) {
-  function Wrapper({ children }) {
-    return <Provider store={store}>{children}</Provider>;
-  }
-  return rtlRender(ui, { wrapper: Wrapper, ...renderOptions });
-}
-
-// re-export everything
-export * from '@testing-library/react';
-// override render method
-export { render };
